Hibernate with H2 database introduced

+++ src/test/java/io/github/mat3e/HelloServiceTest.java
@@ -60,7 +60,7 @@ public class HelloServiceTest {
         //given
         var mockRepository = new LangRepository(){
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 return Optional.empty();
             }
         };
@@ -73,7 +73,7 @@ public class HelloServiceTest {
     private static LangRepository fallbackLangIdRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.FALLBACK_LANG.getId())) {
                     return Optional.of(new Lang(null, FALLBACK_ID_WELCOME, null));
                 }
@@ -84,7 +84,7 @@ public class HelloServiceTest {
     private static LangRepository alwaysReturningHelloRepository() {
         return new LangRepository() {
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 return Optional.of(new Lang(null, WELCOME, null));
             }
         };
